moved ObservableInstanceState out of library
