+++ statime-linux/src/observer.rs
@@ -1,4 +1,7 @@
-use statime::observability::ObservableInstanceState;
+use statime::{
+    config::TimePropertiesDS,
+    observability::{current::CurrentDS, default::DefaultDS, parent::ParentDS},
+};
 use std::{fs::Permissions, os::unix::prelude::PermissionsExt, path::Path, time::Instant};
 use tokio::{io::AsyncWriteExt, net::UnixStream, task::JoinHandle};
 
@@ -6,6 +9,19 @@ use crate::{
     config::Config,
     metrics::exporter::{ObservableState, ProgramData},
 };
+
+/// Observable version of the InstanceState struct
+#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize)]
+pub struct ObservableInstanceState {
+    /// A concrete implementation of the PTP Default dataset (IEEE1588-2019 section 8.2.1)
+    pub default_ds: DefaultDS,
+    /// A concrete implementation of the PTP Current dataset (IEEE1588-2019 section 8.2.2)
+    pub current_ds: CurrentDS,
+    /// A concrete implementation of the PTP Parent dataset (IEEE1588-2019 section 8.2.3)
+    pub parent_ds: ParentDS,
+    /// A concrete implementation of the PTP Time Properties dataset (IEEE1588-2019 section 8.2.4)
+    pub time_properties_ds: TimePropertiesDS,
+}
 
 pub async fn spawn(
     config: &Config,
